Export mesh/empty custom properties to "metadata".
